Add files via upload
Problem Statement

"Blooming Petals" is selling various types of flowers and floral decorations. The company advertises by distributing the pamphlets, advertising on television and so on. Due to rapid development in internet field, the company decides to launch a website where people will get all the information about the various products available to them easily.

Requirement Specification:

The Web site is to be created based on the following requirements:

The Home Page should be created with sections, graphics and with a suitable logo.

Home Page should contain details of different flowers/floral decors with suitable images.

The site should display a menu which will contain the options for brief introduction about the various floral themes/decorations available, location of the shop and any other information if required.

The information about products should be presented in two ways: (a) Information should be available category wise, wherein user can click on a particular category and should be able to view all the products related to the category available with complete information. Eg: Category can be "Birthdays", "Anniversaries", "Marriages" etc. with suitable sub-categories being defined.

(b) Information should be categorized according to the type of flowers like if a User wants to see only "Rose decors" or any other particular flower products then he/she can click on a Link/button/menu etc and can see only that products.

When a user selects any particular flower, a list of products for that brand will be displayed.

A brief summary of individual products should be displayed on the Web Page along with the product but detailed Features should be stored in Individual Word documents which can be downloaded or viewed by the User who wishes to see the same.

The user should also be able to compare the various products of different as well as similar categories.

The Web Page has a "Contact Us" page which will have the Address of the company along with the mail address while using Geolocation API (eg. GoogleMaps) and the mail address when clicked will invoke the local mail client from where they can send an email.